Add Global Exeptions, BG Service and other improvements. Also fixed several bugs and improved overall codebase

+++ DoitFinal.Forum/Controllers/CommentsController.cs
@@ -8,26 +8,30 @@ using System.Security.Claims;
 public class CommentsController : ControllerBase
 {
     private readonly CommentService _commentService;
+    private readonly TopicService _topicService;
 
-    public CommentsController(CommentService commentService)
+    public CommentsController(CommentService commentService, TopicService topicService)
     {
-        _commentService = commentService;
+        _commentService = commentService ?? throw new ArgumentNullException(nameof(commentService));
+        _topicService = topicService ?? throw new ArgumentNullException(nameof(topicService));
     }
 
-    [HttpGet]
-    public async Task<ActionResult<ApiResponse>> Get()
+    [HttpGet("topic/{id}")]
+    public async Task<ActionResult<ApiResponse>> GetAllCommentsByTopic(int id)
     {
-        var commentDTOs = await _commentService.GetAllCommentsAsync();
-        return Ok(CreateApiResponse(commentDTOs, 200, true, "Comments fetched successfully"));
+        var topicComments = await _topicService.GetTopicByIdAsync(id);
+
+        var comments = topicComments.Comments.ToList();
+        return Ok(CreateApiResponse(comments, 200, true, "Comments fetched successfully"));
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<ApiResponse>> Get(int id)
+    public async Task<ActionResult<ApiResponse>> GetCommentById(int id)
     {
         try
         {
-            var commentDTO = await _commentService.GetCommentByIdAsync(id);
-            return Ok(CreateApiResponse(commentDTO, 200, true, "Comment fetched successfully"));
+            var comment = await _commentService.GetCommentByIdAsync(id);
+            return Ok(CreateApiResponse(comment, 200, true, "Comment fetched successfully"));
         }
         catch (Exception ex)
         {
@@ -37,23 +41,43 @@ public class CommentsController : ControllerBase
 
     [HttpPost]
     [Authorize]
-    public async Task<ActionResult<ApiResponse>> Post([FromBody] CommentDTO commentDTO)
+    public async Task<ActionResult<ApiResponse>> CreateComment([FromBody] CommentDTO commentDTO)
     {
         if (!User.Identity.IsAuthenticated)
         {
             return Unauthorized();
         }
 
+        var topic = await _topicService.GetTopicByIdAsync(commentDTO.TopicId);
+        if (topic.Status == TopicStatus.Inactive)
+        {
+            return BadRequest(CreateApiResponse(null, 400, false, "Topic is inactive and comments cannot be added"));
+        }
+
         string userId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
         var createdCommentDTO = await _commentService.CreateCommentAsync(commentDTO, userId);
-        return CreatedAtAction(nameof(Get), CreateApiResponse(createdCommentDTO, 201, true, "Comment created successfully"));
+        return CreatedAtAction(nameof(GetCommentById), CreateApiResponse(createdCommentDTO, 201, true, "Comment created successfully"));
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult<ApiResponse>> Put(int id, [FromBody] CommentDTO commentDTO)
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> UpdateComment(int id, [FromBody] CommentDTO commentDTO)
     {
         try
         {
+            var comment = await _commentService.GetCommentByIdAsync(id);
+            var topic = await _topicService.GetTopicByIdAsync(comment.TopicId);
+            if (topic.Status == TopicStatus.Inactive)
+            {
+                return BadRequest(CreateApiResponse(null, 400, false, "Topic is inactive and comments cannot be updated"));
+            }
+
+            string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+            if (comment.UserEmail != currentUserEmail)
+            {
+                return Unauthorized();
+            }
+
             await _commentService.UpdateCommentAsync(commentDTO);
             return Ok(CreateApiResponse(commentDTO, 200, true, "Comment updated successfully"));
         }
@@ -64,8 +88,17 @@ public class CommentsController : ControllerBase
     }
 
     [HttpDelete("{id}")]
-    public async Task<ActionResult<ApiResponse>> Delete(int id)
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> DeleteComment(int id)
     {
+
+        var comment = await _commentService.GetCommentByIdAsync(id);
+        string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+        if (comment.UserEmail != currentUserEmail)
+        {
+            return Unauthorized();
+        }
+
         await _commentService.DeleteCommentAsync(id);
         return NoContent();
     }
