+++ DoitFinal.Forum/Controllers/TopicsController.cs
@@ -11,18 +11,18 @@ public class TopicsController : ControllerBase
 
     public TopicsController(TopicService topicService)
     {
-        _topicService = topicService;
+        _topicService = topicService ?? throw new ArgumentNullException(nameof(topicService));
     }
 
     [HttpGet]
-    public async Task<ActionResult<ApiResponse>> Get()
+    public async Task<ActionResult<ApiResponse>> GetAllTopics()
     {
         var topicDTOs = await _topicService.GetAllTopicsAsync();
         return Ok(CreateApiResponse(topicDTOs, 200, true, "Topics fetched successfully"));
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<ApiResponse>> Get(int id)
+    public async Task<ActionResult<ApiResponse>> GetTopicById(int id)
     {
         try
         {
@@ -37,7 +37,7 @@ public class TopicsController : ControllerBase
 
     [HttpPost]
     [Authorize]
-    public async Task<ActionResult<ApiResponse>> Post([FromBody] TopicDTO topicDTO)
+    public async Task<ActionResult<ApiResponse>> CreateTopic([FromBody] TopicDTO topicDTO)
     {
         if (!User.Identity.IsAuthenticated)
         {
@@ -47,14 +47,26 @@ public class TopicsController : ControllerBase
         string userEmail = User.Identity.Name;
         string UserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
         var createdTopicDTO = await _topicService.CreateTopicAsync(topicDTO, userEmail, UserId);
-        return CreatedAtAction(nameof(Get), CreateApiResponse(createdTopicDTO, 201, true, "Topic created successfully"));
+        return CreatedAtAction(nameof(GetTopicById), CreateApiResponse(createdTopicDTO, 201, true, "Topic created successfully"));
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult<ApiResponse>> Put(int id, [FromBody] TopicDTO topicDTO)
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> UpdateTopic(int id, [FromBody] TopicDTO topicDTO)
     {
         try
         {
+            var topic = await _topicService.GetTopicByIdAsync(id);
+            if (topic.Status == TopicStatus.Inactive)
+            {
+                return BadRequest(CreateApiResponse(null, 400, false, "Topic is inactive and cannot be updated"));
+            }
+            string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+            if (topic.UserEmail != currentUserEmail)
+            {
+                return Unauthorized();
+            }
+
             await _topicService.UpdateTopicAsync(id, topicDTO);
             return Ok(CreateApiResponse(null, 200, true, "Topic updated successfully"));
         }
@@ -64,10 +76,61 @@ public class TopicsController : ControllerBase
         }
     }
 
+    [HttpPatch("{id}/state")]
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> UpdateTopicState(int id, [FromBody] TopicState state)
+    {
+        try
+        {
+            var topic = await _topicService.GetTopicByIdAsync(id);
+            string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+            if (topic.UserEmail != currentUserEmail)
+            {
+                return Unauthorized();
+            }
+
+            await _topicService.UpdateTopicStateAsync(id, state);
+            return Ok(CreateApiResponse(null, 200, true, "Topic state updated successfully"));
+        }
+        catch (Exception ex)
+        {
+            return BadRequest(CreateApiResponse(null, 400, false, ex.Message));
+        }
+    }
+
+    [HttpPatch("{TopicId}/status")]
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> UpdateTopicStatus(int TopicId, [FromBody] TopicStatus status)
+    {
+        try
+        {
+            var topic = await _topicService.GetTopicByIdAsync(TopicId);
+            string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+            if (topic.UserEmail != currentUserEmail)
+            {
+                return Unauthorized();
+            }
+
+            await _topicService.UpdateTopicStatus(TopicId, status);
+            return Ok(CreateApiResponse(null, 200, true, "Topic status updated successfully"));
+        }
+        catch (Exception ex)
+        {
+            return BadRequest(CreateApiResponse(null, 400, false, ex.Message));
+        }
+    }
 
     [HttpDelete("{id}")]
-    public async Task<ActionResult<ApiResponse>> Delete(int id)
+    [Authorize]
+    public async Task<ActionResult<ApiResponse>> DeleteTopic(int id)
     {
+        var topic = await _topicService.GetTopicByIdAsync(id);
+        string currentUserEmail = User.FindFirst(ClaimTypes.Email).Value;
+        if (topic.UserEmail != currentUserEmail)
+        {
+            return Unauthorized();
+        }
+
         await _topicService.DeleteTopicAsync(id);
         return NoContent();
     }
